feat(java): tema parcialmente implementado.

+++ M8_Java/lista2/src/main/java/br/cwi/crescer/lista2/FileUtilsImpl.java
@@ -7,6 +7,8 @@ package br.cwi.crescer.lista2;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  *
@@ -16,13 +18,24 @@ public class FileUtilsImpl implements FileUtils {
 
     public boolean mk(String string) {
         try{
-            return new File(string).createNewFile();
+            File file = new File(string);
+            if (file.getParentFile() != null) {
+                file.getParentFile().mkdirs();
+            }
+            return file.exists() || string.matches(".*\\..{3}") ? file.createNewFile() : file.mkdir();
         }
-        catch (IOException e){ return false; }
+        catch (IOException e){ 
+            Logger.getLogger(FileUtilsImpl.class.getName()).log(Level.SEVERE, null, e);
+            return false; 
+        }
     }
 
     public boolean rm(String string) {
         File file = new File (string);
+        if(!file.exists())
+            throw new RuntimeException("Arquivo não existe.");
+        if(file.isDirectory())
+            throw new RuntimeException("Arquivo é inválido");
         return file.delete();
     }
 
@@ -38,14 +51,12 @@ public class FileUtilsImpl implements FileUtils {
     }
 
     public boolean mv(String in, String out) {
-        try {
-            File file = new File(in);
-            if (file.isDirectory()){
-                return false;
-            }
-            return file.renameTo(new File(out));
+        File file = new File(in);
+        File fileOut = new File (out);
+        if (file.isDirectory() || fileOut.isDirectory()){
+            throw new RuntimeException("Arquivo é inválido");
         }
-        catch (Exception e) { return false; }
+        return file.renameTo(new File(out));
     }
     
 }
